Agregar middlewares a las rutas de propiedades, clientes, agentes y usuarios. Nueva ruta para conseguir las propiedades del agente logeado y borrar una propiedad requiere ser el agente dueño de la propiedad

+++ src/controllers/propertyController.js
@@ -1,4 +1,6 @@
 const Property = require('../models/property')
+const Agent = require('../models/agent')
+const jwt = require("jsonwebtoken");
 
 const createProperty = async (req, res) => {
   try {
@@ -21,7 +23,6 @@ const getProperties = async (req, res) => {
   }
 };
 
-
 const getPropertyById = async (req, res) => {
   try {
     const { id } = req.params;
@@ -38,36 +39,67 @@ const getPropertyById = async (req, res) => {
   }
 };
 
-const getPropertyByAgent = async (req, res) =>{
+const getPropertiesByAgent = async (req, res) =>{
   try{
-    const { agentId } = req.params;
-    const user = await Property.findAll({
+    const headerAuth = req.headers['authorization']
+    const payload = jwt.verify(headerAuth, process.env.JWT_SECRET);
+    const userId = payload.id
+
+    const agent = await Agent.findAll({
       where: {
-        email: email
+        user: userId
       }
     })
-  }catch (error){
 
+    const agentId = agent[0].id
+
+    const properties = await Property.findAll({
+      where: {
+        agent: agentId
+      }
+    })
+
+    res.status(200).json(properties);
+  }catch (error){
+    console.error(error); // Imprime el error en la consola
+    res.status(500).json({ error: 'Error obteniendo las propiedades' });
   }
 }
 
 const deleteProperty = async (req, res) => {
   try {
-    const { id } = req.params;
-    Property.destroy({
+    const headerAuth = req.headers['authorization']
+    const payload = jwt.verify(headerAuth, process.env.JWT_SECRET);
+    const userId = payload.id
+
+    const agent = await Agent.findAll({
       where: {
-        id: id
+        user: userId
       }
-    }).then(() => {
-      return res.status(200).json({response: 'Propiedad borrada'})
-    }).catch((error) => {
-      res.status(500).json({ error: error });
-    });
+    })
+
+    const agentId = agent[0].id
+    const { propertyId } = req.params;
+    const property = await Property.findByPk(propertyId)
+    if (agentId == property.agent){
+      Property.destroy({
+        where: {
+          id: propertyId
+        }
+      }).then(() => {
+        return res.status(200).json({response: 'Propiedad borrada'})
+      }).catch((error) => {
+        res.status(500).json({ error: error });
+      });
+    }
+    else{
+      throw 'No es dueño de la propiedad'
+    }
 
   } catch (error) {
     console.error(error);
-    res.status(500).json({ error: 'Error obteniendo la propiedad' });
+    res.status(500).json({ error: error });
   }
 };
 
-module.exports = { createProperty, getProperties, getPropertyById, deleteProperty };
+module.exports = { createProperty, getProperties, getPropertyById, getPropertiesByAgent, deleteProperty };
